created several page.js files and component.js files and fitlered them through the react-router-dom router generation
