+++ client/src/Components/Footer.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Footer = () => {
+    return (
+        <div>
+            Footer
+        </div>
+    )
+}
+
+export default Footer;
