+++ client/src/Components/Navbar.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+let Navbar = () => {
+    return (
+        <div>
+            NavBar
+        </div>
+    )
+}
+
+export default Navbar;
