--- client/src/Main/Main.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
- const Main = () => {
-    return (
-       <div>
-           This is the main page vertical landing page. 
-       </div>
-    )
-}
-
-export default Main;
