+++ client/src/Pages/AreaAttractions.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const AreaAttractions = () => {
+    return (
+        <div>
+            AreaAttractions
+        </div>
+    )
+}
+
+export default AreaAttractions;
