+++ client/src/Pages/Dining.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Dining = () => {
+    return (
+        <div>
+            Dining
+        </div>
+    )
+}
+
+export default Dining;
