+++ client/src/Pages/Main.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+ const Main = () => {
+    return (
+       <div>
+           This is the main vertical landing page. 
+       </div>
+    )
+}
+
+export default Main;
