+++ client/src/Pages/Rooms.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Rooms = () => {
+    return(
+        <div>
+            Rooms
+        </div>
+    )
+}
+
+export default Rooms;
