+++ client/src/Pages/Spa.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Spa = () => {
+    return (
+        <div>
+            Spa
+        </div>
+    )
+}
+
+export default Spa;
